refactor(Services): :recycle: Rename addresses folder
