+++ src/config/routes.ts
@@ -1,6 +1,6 @@
-import { ServicesPaths } from '@pages/Addresses/Services.constants';
 import { ProcessesPaths } from '@pages/Processes/Processes.constants';
 import { ProcessGroupsPaths } from '@pages/ProcessGroups/ProcessGroups.constants';
+import { ServicesPaths } from '@pages/Services/Services.constants';
 import { SitesPaths } from '@pages/Sites/Sites.constants';
 import { TopologyPaths } from '@pages/Topology/Topology.constants';
