+++ src/routes.tsx
@@ -2,9 +2,9 @@ import { lazy } from 'react';
 
 import { RouteObject } from 'react-router-dom';
 
-import { servicesRoutes } from '@pages/Addresses/routes';
 import { processesRoutes } from '@pages/Processes/routes';
 import { processGroupsRoutes } from '@pages/ProcessGroups/routes';
+import { servicesRoutes } from '@pages/Services/routes';
 import { errorsRoutes } from '@pages/shared/Errors/routes';
 import { siteRoutes } from '@pages/Sites/routes';
 import { topologyRoutes } from '@pages/Topology/routes';
